Дз 7 модуль

+++ IntOrInteger/src/Container.java
@@ -1,7 +1,7 @@
 public class Container {
-    private Integer count;
+    private int count;
 
-    public void addCount(double value) {
+    public void addCount(int value) {
         count = count + value;
     }
